Made changes to blogsy_controller, transferring those routes to their respective controllers. Will be deleting blogsy_controller on next commit.
