+++ app/models/post.rb
@@ -2,7 +2,7 @@ class Post < ApplicationRecord
 
   belongs_to :user, counter_cache: true
 
-  has_many :comments
+  has_many :comments, dependent: :destroy
 
   validates :user, :title, :body, presence: true
